requirements.txt
fastapi==0.61.1
gunicorn==20.0.4
pandas==1.1.2
psycopg2-binary==2.8.6
selenium==3.141.0
uvicorn==0.11.8
